Dropping need for react for now, json schema load method and reorganize rules.js
